프로젝트 인설트 완료
헤더에 부재 추가 완료

+++ sevenFS/src/main/java/kr/or/ddit/sevenfs/vo/project/ProjectTaskVO.java
@@ -10,24 +10,27 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
 @Data
-@JsonIgnoreProperties(ignoreUnknown = true)
+/*
+ * VO에 없는 필드가 있음 chargerEmpNm, parentTaskNm, parentIndex jaxkson이 오류 없이 무시할 수 있음
+ */
+@JsonIgnoreProperties(ignoreUnknown = true)  
 public class ProjectTaskVO {
-	private int progrsrt;
-	private String taskSttus;
-	private int atchFileNo;
-	private Date taskBeginDt;
-	private int taskDaycnt;
-	private Date taskEndDt;
-	private int taskNo;
-	private long prjctNo;
-	private long upperTaskNo;
-	private String chargerEmpno;
-	private String taskNm;
-	private String taskCn;
-	private String priort;
-	private String taskGrad;
-	
-	private List<MultipartFile> files;
+    private int progrsrt;
+    private String taskSttus;
+    private long atchFileNo;
+    private Date taskBeginDt;
+    private int taskDaycnt;
+    private Date taskEndDt;
+    private int taskNo;
+    private long prjctNo;
+    private Long upperTaskNo;
+    private String chargerEmpno;
+    private String taskNm;
+    private String taskCn;
+    private String priort;
+    private String taskGrad;
+
+    private List<MultipartFile> files;
 
 
 }
